GHC.TypeLits.Printf

Description

Master table comparing the different methods:

Method 	 True Polyarity 	 Naked Arguments 	 Type feedback

pprintf 	 Yes 	 No (requires PP) 	 Yes

rprintf 	 No (HList-based) 	 Yes 	 Yes

printf 	 Yes 	 Yes 	 No

Formattable things

class FormatChar (t :: SChar) a where

Typeclass associating format types (d, f, etc.) with the types that can be formatted by them.

You can extend the printf methods here for your own types by writing your instances here.

Minimal complete definition

Nothing

Methods

formatArg :: p t -> a -> FieldFormat -> ShowS

default formatArg :: PrintfArg a => p t -> a -> FieldFormat -> ShowS

type SChar = Symbol

A type synonym for a single-character symbol

Guarded polyarity

pprintf :: forall str ps. (RPrintf str ps, RecordCurry ps) => CurriedF PP ps String

Type-safe printf with true guarded polyarity. Call it like pprintf @"you have %.02f dollars, %s".

A call to printf on a valid string will always give a well-defined type for a function in return:

>>> :t pprintf @"You have %.2f dollars, %s"
PP "f" -> PP "s" -> String

You can always query the type, and get a well-defined type back, which you can utilize using typed holes or other type-guided development techniques.

To give pprintf its arguments, however, they must be wrapped in PP:

>>> putStrLn $ pprintf @"You have %.2f dollars, %s" (PP 3.62) (PP "Luigi")
You have 3.62 dollars, Luigi

See printf for a polyariadic method that doesn't require PP on its inputs, but doesn't have as usable a type when queried before supplying its arguments, and rprintf for a fake-polyariadic method that doesn't require PP, but requires arguments in a single list instead.

pprintf_ :: forall str ps p. (RPrintf str ps, RecordCurry ps) => p str -> CurriedF PP ps String

data PP (c :: SChar)

Required wrapper around inputs to pprintf (guarded polyarity). See documentation for pprintf for examples of usage.

You can "wrap" any value in PP as long as it can be formatted as the format type indicated.

For example, to make a PP "f", you can use PP 3.5 or PP 94.2, but not PP (3 :: Int) or PP "hello". To make a value of type PP c, you must wrap a value that can be formatted via c.

Constructors

forall a.FormatChar c a => PP a

List-based polyarity

data Rec (a :: u -> Type) (b :: [u]) where

A record is parameterized by a universe u, an interpretation f and a list of rows rs. The labels or indices of the record are given by inhabitants of the kind u; the type of values at any label r :: u is given by its interpretation f r :: *.

Constructors

RNil :: forall u (a :: u -> Type). Rec a ('[] :: [u])

Bundled Patterns

pattern (:%) :: () => FormatChar c a => a -> FormatArgs cs -> FormatArgs (c ': cs) infixr 7

type FormatArgs = Rec PP

rprintf :: forall str ps. RPrintf str ps => FormatArgs ps -> String

Type-safe printf with faked polyarity. Pass in a "list" of arguments (using :% and RNil), instead of as multiple arguments. Call it like rprintf @"you have %.02f dollars, %s".

>>> :t rprintf @"You have %.2f dollars, %s"
FormatArgs '["f", "s"] -> String

This means that it is expecting something that can be printed with f and something that can be printed with s. We can provide a Double and a String:

>>> putStrLn $ rprintf @"You have %.2f dollars, %s" (3.62 ':%' "Luigi" :% 'RNil')
You have 3.62 dollars, Luigi

See pprintf for a version with true polyarity and good clear types, but requires wrapping its arguments, and printf for a version with true polyarity but less clear types.

rprintf_ :: RPrintf str ps => p str -> FormatArgs ps -> String

Unguarded polyarity

printf :: forall str fun. Printf str fun => fun

Type-safe printf with true naked polyarity. Call it like printf @"you have %.02f dollars, %s".

>>> putStrLn $ printf @"You have %.2f dollars, %s" 3.62 "Luigi"
You have 3.62 dollars, Luigi

If you what the type of this function unapplied (or with not enough arguments), or try to use it with typed holes or type-guided development, the type errors aren't going to be pretty in most situations. In addition, you always have to make sure the result type can be inferred as String, which may require a type annotation in some situations.

(Measures have been taken to make the error messages as helpful as possible, but they're not going to be as pretty as for pprintf or rprintf)

However, if you use it properly with the right number of arguments, everything should work and be type-safe: code with missing or badly typed arguments will not compile.

See pprintf for a version of this with much nicer types and type errors, but requires wrapping arguments, and rprintf for a version of this with "fake" polyarity, taking a list as input instead.

printf_ :: Printf str fun => p str -> fun

class FormatFun (ffs :: [Either Symbol FieldFormat]) fun

Minimal complete definition

formatFun