GHC.TypeLits.Printf.Internal

Documentation

class FormatChar (t :: SChar) a where

Typeclass associating format types (d, f, etc.) with the types that can be formatted by them.

You can extend the printf methods here for your own types by writing your instances here.

Minimal complete definition

Nothing

Methods

formatArg :: p t -> a -> FieldFormat -> ShowS

default formatArg :: PrintfArg a => p t -> a -> FieldFormat -> ShowS

data PP (c :: SChar)

Required wrapper around inputs to pprintf (guarded polyarity). See documentation for pprintf for examples of usage.

You can "wrap" any value in PP as long as it can be formatted as the format type indicated.

For example, to make a PP "f", you can use PP 3.5 or PP 94.2, but not PP (3 :: Int) or PP "hello". To make a value of type PP c, you must wrap a value that can be formatted via c.

Constructors

forall a.FormatChar c a => PP a

class RPrintf (str :: Symbol) ps where

Methods

rprintf_ :: p str -> FormatArgs ps -> String

A version of rprintf taking an explicit proxy, which allows usage without TypeApplications

>>> :t rprintf_ (Proxy :: Proxy "You have %.2f dollars, %s")
FormatArgs '["f", "s"] -> String

type FormatArgs = Rec PP

class RFormat (ffs :: [Either Symbol FieldFormat]) (ps :: [SChar]) | ffs -> ps where

Methods

rformat :: p ffs -> FormatArgs ps -> ShowS

class Printf (str :: Symbol) fun where

Methods

printf_ :: p str -> fun

A version of printf taking an explicit proxy, which allows usage without TypeApplications

>>> putStrLn $ printf_ (Proxy :: Proxy "You have %.2f dollars, %s") 3.62 "Luigi"
You have 3.62 dollars, Luigi

class FormatFun (ffs :: [Either Symbol FieldFormat]) fun where

Methods

formatFun :: p ffs -> String -> fun