GHC.TypeLits.Printf

Description

An extensible and type-safe printf from parsing GHC TypeLits Symbol literals, matching the semantics of printf from Text.Printf in base. The difference is that the variants here will always fail to compile if given arguments of the wrong type (or too many or too little arguments). Most of the variants also provide useful type feedback, telling you the type of arguments it expects and how many when queried with :t or with typed holes. See documentation in Text.Printf for details on how this formats items of various types, and the differences with C printf(3).

Comparing their usage/calling conventions:

>>> putStrLn $ pprintf @"You have %.2f dollars, %s" (PP 3.62) (PP "Luigi")
You have 3.62 dollars, Luigi
>>> putStrLn $ rprintf @"You have %.2f dollars, %s" (3.62 :% "Luigi" :% RNil)
You have 3.62 dollars, Luigi
>>> putStrLn $ printf @"You have %.2f dollars, %s" 3.62 "Luigi"
You have 3.62 dollars, Luigi

Now comparing their types:

>>> :t pprintf @"You have %.2f dollars, %s" 3.62 "Luigi"
PP "f" -> PP "s" -> String
>>> :t rprintf @"You have %.2f dollars, %s" 3.62 "Luigi"
FormatArgs '["f", "s"] -> String
>>> :t printf @"You have %.2f dollars, %s" 3.62 "Luigi"
FormatFun '[ .... ] fun => fun

For pprintf, it shows you need two arguments: A PP "f" (which is a value that supports being formatted by f) like PP 3.62, and a PP "s", like PP Luigi.

rprintf tells you you need a two-item hlist (from Data.Vinyl.Core), where the first item implements f and the second item implements s: 3.62 :% Luigi :% RNil will do.

The type of printf is much less informative. It's possible to see what you need from the ... in FormatFun...but it's basically a situation that works fine when it does, but can be tricky if you mess up. The up-side is that it is the cleanest to call if you already know what you need: you can just give the arguments plainly, like 3.62 and Luigi.

The type of printf doesn't tell you immediately what you you need. However, if you do try to use it, the type errors will guide you along the way, iteratively.

>>> printf @"You have %.2f dollars, %s"
-- ERROR: Call to printf missing argument fulfilling "%.2f"
-- Either provide an argument or rewrite the format string to not expect
-- one.

>>> printf @"You have %.2f dollars, %s" 3.62
-- ERROR: Call to printf missing argument fulfilling "%s"
-- Either provide an argument or rewrite the format string to not expect
-- one.

>>> printf @"You have %.2f dollars, %s" 3.62 "Luigi"
You have 3.62 dollars, Luigi

>>> printf @"You have %.2f dollars, %s" 3.62 "Luigi" 72
-- ERROR: An extra argument of type Integer was given to a call to printf
-- Either remove the argument, or rewrite the format string to include the
-- appropriate hole.

The following table summarizes the features and drawbacks of each method:

Method 	 True Polyarity 	 Naked Arguments 	 Type feedback

pprintf 	 Yes 	 No (requires PP) 	 Yes

rprintf 	 No (HList-based) 	 Yes 	 Yes

printf 	 Yes 	 Yes 	 Partial (via errors)

Ideally we would have a solution that has all three. However, as of now, we have a "pick two" sort of situation. Suggestions are definitely welcome, however, if you find something that satisfies all three benefits while still allowing for polymorphism!

You can extend functionality with formatting for your own types by providing instances of FormatType.

Also in this module is pfmt, which allows you to format individual items according to a single format specifier.

Formattable things

class FormatType (t :: SChar) a where

Typeclass associating format types (d, f, etc.) with the types that can be formatted by them.

You can extend the printf methods here for your own types by writing your instances here.

Minimal complete definition

Nothing

Methods

formatArg :: p t -> a -> FieldFormat -> ShowS

default formatArg :: PrintfArg a => p t -> a -> FieldFormat -> ShowS

type SChar = Symbol

A type synonym for a single-character symbol. Ideally this would just be Char, but we don't have chars at the type level. So, if you see SChar in a type signature, it means that it's expected to be a symbol/string with only one single character.

Printf

Guarded polyarity

pprintf :: forall str ps. (RPrintf str ps, RecordCurry ps) => CurriedF PP ps String

Type-safe printf with true guarded polyarity. Call it like pprintf @"you have %.02f dollars, %s".

A call to printf on a valid string will always give a well-defined type for a function in return:

>>> :t pprintf @"You have %.2f dollars, %s"
PP "f" -> PP "s" -> String

You can always query the type, and get a well-defined type back, which you can utilize using typed holes or other type-guided development techniques.

To give pprintf its arguments, however, they must be wrapped in PP:

>>> putStrLn $ pprintf @"You have %.2f dollars, %s" (PP 3.62) (PP "Luigi")
You have 3.62 dollars, Luigi

See printf for a polyariadic method that doesn't require PP on its inputs, but doesn't have as usable a type when queried before supplying its arguments, and rprintf for a fake-polyariadic method that doesn't require PP, but requires arguments in a single list instead. Also see top-level module documentation GHC.TypeLits.Printf for a more comprehensive summary.

pprintf_ :: forall str ps p. (RPrintf str ps, RecordCurry ps) => p str -> CurriedF PP ps String

A version of pprintf taking an explicit proxy, which allows usage without TypeApplications

>>> :t pprintf_ (Proxy :: Proxy "You have %.2f dollars, %s")
PP "f" -> PP "s" -> String

data PP (c :: SChar)

Required wrapper around inputs to pprintf (guarded polyarity). See documentation for pprintf for examples of usage.

You can "wrap" any value in PP as long as it can be formatted as the format type indicated.

For example, to make a PP "f", you can use PP 3.5 or PP 94.2, but not PP (3 :: Int) or PP "hello". To make a value of type PP c, you must wrap a value that can be formatted via c.

Constructors

forall a.FormatType c a => PP a

List-based polyarity

rprintf :: forall str ps. RPrintf str ps => FormatArgs ps -> String

Type-safe printf with faked polyarity. Pass in a "list" of arguments (using :% and RNil), instead of as multiple arguments. Call it like rprintf @"you have %.02f dollars, %s".

>>> :t rprintf @"You have %.2f dollars, %s"
FormatArgs '["f", "s"] -> String

This means that it is expecting something that can be printed with f and something that can be printed with s. We can provide a Double and a String:

>>> putStrLn $ rprintf @"You have %.2f dollars, %s" (3.62 ':%' "Luigi" :% 'RNil')
You have 3.62 dollars, Luigi

See pprintf for a version with true polyarity and good clear types, but requires wrapping its arguments, and printf for a version with true polyarity but less clear types. Also see top-level module documentation GHC.TypeLits.Printf for a more comprehensive summary.

rprintf_ :: RPrintf str ps => p str -> FormatArgs ps -> String

A version of rprintf taking an explicit proxy, which allows usage without TypeApplications

>>> :t rprintf_ (Proxy :: Proxy "You have %.2f dollars, %s")
FormatArgs '["f", "s"] -> String

data Rec (a :: u -> Type) (b :: [u]) where

A record is parameterized by a universe u, an interpretation f and a list of rows rs. The labels or indices of the record are given by inhabitants of the kind u; the type of values at any label r :: u is given by its interpretation f r :: *.

Constructors

RNil :: forall u (a :: u -> Type). Rec a ('[] :: [u])

Bundled Patterns

pattern (:%) :: () => FormatType c a => a -> FormatArgs cs -> FormatArgs (c ': cs) infixr 7
Pattern and constructor allowing you to construct a FormatArgs.

To construct a FormatArgs '["f", "s"], for instance, you need to give a value formattable by f and a value formattable by s, given like a linked list, with :% for cons and RNil for nil.

3.62 :% Luigi :% RNil

(This should evoke the idea of of 3.62 : Luigi : [], even though the latter is not possible in Haskell)

type FormatArgs = Rec PP

A heterogeneous list (from Data.Vinyl.Core) used for calling with rprintf. Instead of supplying the inputs as different arguments, we can gather all the inputs into a single list to give to rprintf.

>>> :t rprintf @"You have %.2f dollars, %s"
FormatArgs '["f", "s"] -> String

To construct a FormatArgs '["f", "s"], you need to give a value formattable by f and a value formattable by s, given like a linked list, with :% for cons and RNil for nil.

>>> putStrLn $ rprintf @"You have %.2f dollars, %s" (3.62 :% "Luigi" :% RNil)
You have 3.62 dollars, Luigi

(This should evoke the idea of of 3.62 : Luigi : [], even though the latter is not possible in Haskell)

Unguarded polyarity

printf :: forall str fun. Printf str fun => fun

Type-safe printf with true naked polyarity. Call it like printf @"you have %.02f dollars, %s".

>>> putStrLn $ printf @"You have %.2f dollars, %s" 3.62 "Luigi"
You have 3.62 dollars, Luigi

If you what the type of this function unapplied (or with not enough arguments), or try to use it with typed holes or type-guided development, the type errors aren't going to be pretty in most situations. In addition, you always have to make sure the result type can be inferred as String, which may require a type annotation in some situations.

(Measures have been taken to make the error messages as helpful as possible, but they're not going to be as pretty as for pprintf or rprintf. See FormatFun documentation for ways to get some limited type feedback from this function.)

Essentially, this is easier to use if you have all the right arguments in place...but more difficult to debug (at compiletime) if you mess up.

Some guidelines for making sure the type-checking and debugging story goes as nicely as possible:

Make sure the result type is always known monomorphically. Sometimes this means requiring an explicit annotation, like printf ... :: String.

Make sure all the values you give to this have known monomorphic types, as well.

However, all debugging should still be only at compile-time. Once it compiles, it's safe --- code with missing or badly typed arguments will not compile, and so won't give you any runtime errors.

See pprintf for a version of this with much nicer types and type errors, but requires wrapping arguments, and rprintf for a version of this with "fake" polyarity, taking a list as input instead. Also see top-level module documentation GHC.TypeLits.Printf for a more comprehensive summary.

Note that this also supports the "interpret as an IO action to print out results" functionality that Text.Printf supports. This also supports returning strict Text and lazy Text as well.

printf_ :: Printf str fun => p str -> fun

A version of printf taking an explicit proxy, which allows usage without TypeApplications

>>> putStrLn $ printf_ (Proxy :: Proxy "You have %.2f dollars, %s") 3.62 "Luigi"
You have 3.62 dollars, Luigi

data PHelp

pHelp :: PHelp -> String

class FormatFun (ffs :: [Either Symbol FieldFormat]) fun

The typeclass supporting polyarity used by printf. It works in mostly the same way as PrintfType from Text.Printf, and similar the same as FormatF.

Ideally, you will never have to run into this typeclass or have to deal with it. It will come up if you ask for the type of printf, or if you give the wrong number or type of arguments to it.

>>> :t printf @"You have %.2f dollars, %s"
FormatFun '[ Right ..., 'Left " dollars ", 'Right ...] fun => fun

Every item in the first argument of FormatFun is a chunk of the formatting string, split between format holes (Right) and string chunks (Left). You can successively "eliminate" them by providing more arguments that implement each hole:

>>> :t printf @"You have %.2f dollars, %s" 3.62
FormatFun '[ Right ...] fun => fun

Until you you finally fill all the holes:

>>> :t printf @"You have %.2f dollars, %s" 3.62 "Luigi"
FormatFun '[] t => t

at which point you may use it as a String or IO (), in the same way that Text.Printf works. We also support using strict Text lazy Text as well.

Note that while it is somewhat possible to "reason with" or get type feedback from this (as shown above), it is nowhere near the level of clarity as pprintf or rprintf. And, if things go wrong (like passing too many or too little arguments, or arguments of the wrong type), the error messages are not going to be as useful. However, there are some measures put into place to make type errors more feasible for debugging, and I'm open for suggestions for improvement methods, as well!

Some guidelines for making sure the type-checking and debugging story goes as nicely as possible:

Make sure the result type is always known monomorphically. Sometimes this means requiring an explicit annotation, like printf ... :: String.

Make sure all the values you give to this have known monomorphic types, as well.

Minimal complete definition

formatFun

Single item

pfmt :: forall c a. FormatType c a => PFmt c -> a -> String

Parse and run a single format hole on a single vale. Can be useful for formatting individual items or for testing your own custom instances of FormatType.

Usually meant to be used with OverloadedLabels:

>>> pfmt #f 3.62
"3.62"

However, current versions of GHC disallow labels that aren't valid identifier names, disallowing things like pfmt #.2f 3.62. While there is an <https://github.com/ghc-proposals/ghc-proposals/blob/master/proposals/0170-unrestricted-overloadedlabels.rst approved proposal> that allows this, if you are using an earlier GHC version, you can get around this using mkPFmt:

>>> pfmt (mkPFmt @".2f") 3.6234124
"3.62"

Ideally we'd want to be able to write

>>> pfmt #.2f 3.6234124
"3.62"

(which should be possible in GHC 8.10+)

Note that the format string does not include the leading %.

data PFmt c

Utility type powering pfmt. See dcumentation for pfmt for more information on usage.

Using OverloadedLabels, you never need to construct this directly can just write #f and a PFmt "f" will be generated. You can also create this using mkPFmt or mkPFmt_, in the situations where OverloadedLabels doesn't work or is not wanted.

mkPFmt :: forall str lst ff f w q m c. (Listify str lst, 'Just ff ~ ParseFmt lst, Reflect ff, ff ~ 'FF f w q m c) => PFmt c

Useful for using pfmt without OverloadedLabels, or also when passing format specifiers that aren't currently allowed with OverloadedLabels until GHC 8.10+ (like #.2f).

>>> pfmt (mkPFmt @".2f") 3.6234124
"3.62"

mkPFmt_ :: forall str lst ff f w q m c p. (Listify str lst, 'Just ff ~ ParseFmt lst, Reflect ff, ff ~ 'FF f w q m c) => p str -> PFmt c

A version of mkPFmt that takes an explicit proxy input.

>>> pfmt (mkPFmt_ (Proxy :: Proxy ".2f") 3.6234124
"3.62"